Fixed error with MCD module and Test bench add a 'button press' for reinitilize machine
There was a mistake in MCD_Calcul.vhd with an if condition that make the state machine fail. Apart from that correction Block_Counter is been modify to just output Max std_logic  when counter surpases the given number. Changes in Block_Register are just to make syncronous reset. The rest of the files have minor changes like initialize internal signal to 0. Next task is to make this test on ghdl